# apps/workspace/generated/http/openapi.bundled.yaml
openapi: 3.1.2
info:
  title: Univer Workspace API
  version: 0.1.0
  summary: Product HTTP interface for the Univer Workspace collaboration example.
  description: |
    This contract exposes product concepts such as users, spaces, Nodes,
    Resources, trash batches, grants, operations, and worktrees. Univer
    Collaboration protocol endpoints are configured separately and are not
    redefined here.
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
  /api/session:
    get:
      tags:
        - Session
      operationId: getSession
      summary: Get the current login session.
      security:
        - {}
        - sessionCookie: []
      responses:
        '200':
          description: Current authenticated or anonymous session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionView'
...
  /api/spaces:
    get:
      tags:
        - Spaces
      operationId: listSpaces
      summary: List Spaces discoverable by the current User.
      responses:
        '200':
          description: Personal Space followed by owned and joined Team Spaces.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/team-spaces:
    post:
      tags:
        - Spaces
      operationId: createTeamSpace
      summary: Create a Team Space owned by the current User.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamSpace'
      responses:
        '201':
          description: Created Team Space.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/spaces/{spaceId}:
    parameters:
      - name: spaceId
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Spaces
      operationId: getSpace
      summary: Get a Space the current User may discover.
      responses:
        '200':
          description: Space and effective capabilities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceView'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Spaces
      operationId: updateSpace
      summary: Update Space settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchSpace'
      responses:
        '200':
          description: Updated Space.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/resources:
    post:
      tags:
        - Resources
      operationId: createResource
      summary: Create a Univer Resource and its Node.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResource'
      responses:
        '200':
          description: Existing completed result for the Idempotency Key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceCreateResponse'
        '201':
          description: Resource created during the initial attempt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceCreateResponse'
        '202':
          description: Resource creation remains pending.
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/resources/{resourceId}:
    parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Resources
      operationId: getResource
      summary: Get Resource and owning Node metadata.
      responses:
        '200':
          description: Current Resource and Node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/unit-resources/{unitId}:
    parameters:
      - name: unitId
        in: path
        required: true
        schema:
          type: string
    get:
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Resources
      operationId: getResourceByUnit
      summary: Resolve Resource and owning Node metadata from a Univer Unit.
      responses:
        '200':
          description: Current Resource and Node accessible to the current reader.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/resources/{resourceId}/open:
    parameters:
      - name: resourceId
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - {}
        - sessionCookie: []
      description: Anonymous readers receive viewer access through Link Sharing or Space public read; inaccessible content returns 404.
      tags:
        - Resources
      operationId: openResource
      summary: Resolve the Unit and editor mode for a Resource.
      responses:
        '200':
          description: Resource open descriptor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceOpenView'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: workspace_session
      description: Opaque HttpOnly login session cookie.
    discordBotApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Shared secret configured on Workspace and the trusted Discord Bot server.
  schemas:
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable diagnostic text.
        field:
          type: string
          description: Input field related to the error, when applicable.
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
    User:
      type: object
      additionalProperties: false
      required:
        - id
        - username
        - displayName
        - avatarUrl
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type:
            - string
            - 'null'
          format: uri
...
    SessionView:
      oneOf:
        - $ref: '#/components/schemas/AnonymousSession'
        - $ref: '#/components/schemas/AuthenticatedSession'
    SpaceView:
      allOf:
        - $ref: '#/components/schemas/SpaceSummary'
        - type: object
          required:
            - accessRole
            - capabilities
          properties:
            accessRole:
              $ref: '#/components/schemas/SpaceRole'
            capabilities:
              $ref: '#/components/schemas/SpaceCapabilities'
      unevaluatedProperties: false
    NodeSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - spaceId
        - parentNodeId
        - name
        - resource
        - hasChildren
        - updatedAt
        - accessRole
        - capabilities
      properties:
        id:
          type: string
        spaceId:
          type: string
        parentNodeId:
          type:
            - string
            - 'null'
        name:
          type: string
        resource:
          oneOf:
            - $ref: '#/components/schemas/ResourceSummary'
            - type: 'null'
        hasChildren:
          type: boolean
        updatedAt:
          type: string
          format: date-time
        accessRole:
          $ref: '#/components/schemas/AccessRole'
        capabilities:
          $ref: '#/components/schemas/NodeCapabilities'
...
    OperationView:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - state
        - createdAt
        - updatedAt
        - result
        - error
      properties:
        id:
          type: string
        kind:
          $ref: '#/components/schemas/OperationKind'
        state:
          $ref: '#/components/schemas/OperationState'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        result:
          type:
            - object
            - 'null'
          additionalProperties: true
        error:
          oneOf:
            - $ref: '#/components/schemas/OperationError'
            - type: 'null'
    ResourceOpenView:
      type: object
      additionalProperties: false
      required:
        - resource
      properties:
        resource:
          oneOf:
            - $ref: '#/components/schemas/OpenUniverResource'
            - $ref: '#/components/schemas/OpenBlobResource'
          discriminator:
            propertyName: kind
            mapping:
              univer: '#/components/schemas/OpenUniverResource'
              blob: '#/components/schemas/OpenBlobResource'
...
    AnonymousSession:
      type: object
      additionalProperties: false
      required:
        - authenticated
        - githubOAuthEnabled
        - discordOAuthEnabled
      properties:
        authenticated:
          const: false
        githubOAuthEnabled:
          type: boolean
        discordOAuthEnabled:
          type: boolean
    ExternalIdentitySummary:
      type: object
      additionalProperties: false
      required:
        - provider
        - providerUsername
      properties:
        provider:
          type: string
          enum:
            - github
            - discord
        providerUsername:
          type: string
    AuthenticationMethods:
      type: object
      additionalProperties: false
      required:
        - password
        - externalIdentities
      properties:
        password:
          type: boolean
        externalIdentities:
          type: array
          items:
            $ref: '#/components/schemas/ExternalIdentitySummary'
    AuthenticatedSession:
      type: object
      additionalProperties: false
      required:
        - authenticated
        - githubOAuthEnabled
        - discordOAuthEnabled
        - user
        - authenticationMethods
      properties:
        authenticated:
          const: true
        githubOAuthEnabled:
          type: boolean
        discordOAuthEnabled:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        authenticationMethods:
          $ref: '#/components/schemas/AuthenticationMethods'
...
    SpaceType:
      type: string
      enum:
        - personal
        - team
    SpaceSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - name
        - publicRead
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/SpaceType'
        name:
          type: string
        publicRead:
          type: boolean
          description: Whether anyone with a link can browse and open this Space read-only.
    SpaceRole:
      type: string
      enum:
        - owner
        - admin
        - editor
        - viewer
    SpaceCapabilities:
      type: object
      additionalProperties: false
      required:
        - browseRoot
        - createAtRoot
        - renameSpace
        - manageMembers
        - viewTrash
      properties:
        browseRoot:
          type: boolean
        createAtRoot:
          type: boolean
        renameSpace:
          type: boolean
        manageMembers:
          type: boolean
        viewTrash:
          type: boolean
    SpaceList:
      type: object
      additionalProperties: false
      required:
        - spaces
      properties:
        spaces:
          type: array
          items:
            $ref: '#/components/schemas/SpaceView'
    CreateTeamSpace:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        publicRead:
          type: boolean
          description: Whether anyone with a link can read the Space without signing in. Omit to keep it private.
    PatchSpace:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        publicRead:
          type: boolean
    UnitType:
      type: string
      enum:
        - sheet
        - doc
        - slide
        - board
        - base
    ResourceCapabilities:
      type: object
      additionalProperties: false
      required:
        - openContent
        - editContent
        - downloadContent
      properties:
        openContent:
          type: boolean
        editContent:
          type: boolean
        downloadContent:
          type: boolean
    UniverResourceSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - unitId
        - unitType
        - capabilities
      properties:
        id:
          type: string
        kind:
          type: string
          const: univer
        unitId:
          type: string
        unitType:
          $ref: '#/components/schemas/UnitType'
        capabilities:
          $ref: '#/components/schemas/ResourceCapabilities'
    BlobAvailability:
      type: string
      enum:
        - ready
        - quarantined
    BlobResourceSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - mediaType
        - byteSize
        - availability
        - capabilities
      properties:
        id:
          type: string
        kind:
          type: string
          const: blob
        mediaType:
          type: string
        byteSize:
          type: integer
          minimum: 0
        availability:
          $ref: '#/components/schemas/BlobAvailability'
        capabilities:
          $ref: '#/components/schemas/ResourceCapabilities'
    ResourceSummary:
      oneOf:
        - $ref: '#/components/schemas/UniverResourceSummary'
        - $ref: '#/components/schemas/BlobResourceSummary'
      discriminator:
        propertyName: kind
        mapping:
          univer: '#/components/schemas/UniverResourceSummary'
          blob: '#/components/schemas/BlobResourceSummary'
    AccessRole:
      type: string
      enum:
        - owner
        - admin
        - editor
        - viewer
    NodeCapabilities:
      type: object
      additionalProperties: false
      required:
        - browseChildren
        - createChildren
        - rename
        - move
        - trash
        - share
      properties:
        browseChildren:
          type: boolean
        createChildren:
          type: boolean
        rename:
          type: boolean
        move:
          type: boolean
        trash:
          type: boolean
        share:
          type: boolean
...
    OperationKind:
      type: string
      enum:
        - createResource
        - createBlobResource
        - replaceBlobContent
        - createWorktree
        - addWorktreeUnit
        - createWorktreeUnit
        - mergeWorktree
        - discardWorktree
        - activateWorktreeResource
    OperationState:
      type: string
      enum:
        - pending
        - completed
        - failed
    OperationError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    CreateResource:
      type: object
      additionalProperties: false
      required:
        - kind
        - spaceId
        - parentNodeId
        - name
        - unitType
      properties:
        kind:
          type: string
          const: univer
        spaceId:
          type: string
        parentNodeId:
          type:
            - string
            - 'null'
        name:
          type: string
          minLength: 1
          maxLength: 255
        unitType:
          $ref: '#/components/schemas/UnitType'
        initialData:
          type: object
          description: Complete initial Univer Unit data for unitType. When omitted, the server creates a blank Unit. The server always overrides id with the Unit identity reserved from the Idempotency-Key and resets rev to the initial revision.
          additionalProperties: true
    ResourceCreateResponse:
      type: object
      additionalProperties: false
      required:
        - operation
        - node
      properties:
        operation:
          $ref: '#/components/schemas/OperationView'
        node:
          $ref: '#/components/schemas/NodeSummary'
    OperationEnvelope:
      type: object
      additionalProperties: false
      required:
        - operation
      properties:
        operation:
          $ref: '#/components/schemas/OperationView'
    ResourceResponse:
      type: object
      additionalProperties: false
      required:
        - resource
        - node
      properties:
        resource:
          $ref: '#/components/schemas/ResourceSummary'
        node:
          $ref: '#/components/schemas/NodeSummary'
    EditorMode:
      type: string
      enum:
        - edit
        - readOnly
    OpenUniverResource:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - nodeId
        - spaceId
        - name
        - unitId
        - unitType
        - accessRole
        - editorMode
      properties:
        id:
          type: string
        kind:
          type: string
          const: univer
        nodeId:
          type: string
        spaceId:
          type: string
        name:
          type: string
        unitId:
          type: string
        unitType:
          $ref: '#/components/schemas/UnitType'
        accessRole:
          $ref: '#/components/schemas/AccessRole'
        editorMode:
          $ref: '#/components/schemas/EditorMode'
    OpenBlobResource:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - nodeId
        - spaceId
        - name
        - accessRole
        - originalFilename
        - mediaType
        - byteSize
        - sha256
        - contentUrl
        - downloadUrl
      properties:
        id:
          type: string
        kind:
          type: string
          const: blob
        nodeId:
          type: string
        spaceId:
          type: string
        name:
          type: string
        accessRole:
          $ref: '#/components/schemas/AccessRole'
        originalFilename:
          type: string
        mediaType:
          type: string
        byteSize:
          type: integer
          minimum: 0
        sha256:
          type: string
        contentUrl:
          type: string
        downloadUrl:
          type: string
...
  parameters:
...
    IdempotencyKeyParameter:
      name: Idempotency-Key
      in: header
      required: true
      description: Random opaque key representing one user intent.
      schema:
        type: string
        minLength: 16
        maxLength: 200
        pattern: ^[A-Za-z0-9_-]+$
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: A valid login session is required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthenticated:
              value:
                error:
                  code: UNAUTHENTICATED
                  message: Authentication is required.
    Forbidden:
      description: The current User cannot perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The resource does not exist or is not discoverable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: NOT_FOUND
                  message: The resource was not found.
    Conflict:
      description: The request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...